Fikser testdata for lokal kjøring
