+++ server/src/local/aggregertStatistikkMockdata.ts
@@ -1,0 +1,137 @@
+export const mockdataOrgnr91096939 = {
+  prosentSiste4Kvartaler: [
+    {
+      statistikkategori: "LAND",
+      label: "Norge",
+      verdi: "9.0",
+      antallPersonerIBeregningen: 10,
+      kvartalerIBeregningen: [
+        {
+          årstall: 2022,
+          kvartal: 1,
+        },
+      ],
+    },
+    {
+      statistikkategori: "NÆRING",
+      label: "Barnehagenæringen",
+      verdi: "10.0",
+      antallPersonerIBeregningen: 20000,
+      kvartalerIBeregningen: [
+        {
+          årstall: 2022,
+          kvartal: 1,
+        },
+      ],
+    },
+    {
+      statistikkategori: "BRANSJE",
+      label: "Barnehager",
+      verdi: "19.0",
+      antallPersonerIBeregningen: 20000,
+      kvartalerIBeregningen: [
+        {
+          årstall: 2022,
+          kvartal: 1,
+        },
+      ],
+    },
+  ],
+  trend: [
+    {
+      statistikkategori: "NÆRING",
+      label: "Barnehagenæringen",
+      verdi: "-2.0",
+      antallPersonerIBeregningen: 20000,
+      kvartalerIBeregningen: [
+        {
+          årstall: 2022,
+          kvartal: 1,
+        },
+        {
+          årstall: 2021,
+          kvartal: 1,
+        },
+      ],
+    },
+    {
+      statistikkategori: "BRANSJE",
+      label: "Barnehager",
+      verdi: "-2.0",
+      antallPersonerIBeregningen: 20000,
+      kvartalerIBeregningen: [
+        {
+          årstall: 2022,
+          kvartal: 1,
+        },
+        {
+          årstall: 2021,
+          kvartal: 1,
+        },
+      ],
+    },
+  ],
+};
+
+export const mockdataOrgnr81096939 = {
+  prosentSiste4Kvartaler: [
+    {
+      statistikkategori: "LAND",
+      label: "Norge",
+      verdi: "9.0",
+      antallPersonerIBeregningen: 10,
+      kvartalerIBeregningen: [
+        {
+          årstall: 2022,
+          kvartal: 1,
+        },
+      ],
+    },
+    {
+      statistikkategori: "NÆRING",
+      label: "Barnehagenæringen",
+      verdi: "-1.0",
+      antallPersonerIBeregningen: 10,
+      kvartalerIBeregningen: [
+        {
+          årstall: 2022,
+          kvartal: 1,
+        },
+      ],
+    },
+  ],
+  trend: [
+    {
+      statistikkategori: "BRANSJE",
+      label: "Barnehager",
+      verdi: "-2.0",
+      antallPersonerIBeregningen: 10,
+      kvartalerIBeregningen: [
+        {
+          årstall: 2022,
+          kvartal: 1,
+        },
+        {
+          årstall: 2021,
+          kvartal: 1,
+        },
+      ],
+    },
+    {
+      statistikkategori: "NÆRING",
+      label: "Barnehagenæringen",
+      verdi: "-1.0",
+      antallPersonerIBeregningen: 10,
+      kvartalerIBeregningen: [
+        {
+          årstall: 2022,
+          kvartal: 1,
+        },
+        {
+          årstall: 2021,
+          kvartal: 1,
+        },
+      ],
+    },
+  ],
+};
