+++ server/src/local/organisasajonerMockdata.ts
@@ -1,0 +1,58 @@
+export const organisasjoner = [
+  {
+    Name: "FLESK OG FISK AS [Local server]",
+    Type: "Enterprise",
+    OrganizationNumber: "111111111",
+    OrganizationForm: "AS",
+    Status: "Active",
+    ParentOrganizationNumber: "",
+  },
+  {
+    Name: "FLESK OG FISK OSLO [Local server]",
+    Type: "Business",
+    OrganizationNumber: "910969439",
+    OrganizationForm: "BEDR",
+    Status: "Active",
+    ParentOrganizationNumber: "111111111",
+  },
+  {
+    Name: "Trøndelag Tømmerere [Local server]",
+    Type: "Enterprise",
+    OrganizationNumber: "211111111",
+    OrganizationForm: "AS",
+    Status: "Active",
+    ParentOrganizationNumber: "",
+  },
+  {
+    Name: "Trøndelag Tømmerere avd. OSLO [Local server]",
+    Type: "Business",
+    OrganizationNumber: "810969439",
+    OrganizationForm: "BEDR",
+    Status: "Active",
+    ParentOrganizationNumber: "211111111",
+  },
+  {
+    Name: "System feil AS",
+    Type: "Enterprise",
+    OrganizationNumber: "311111111",
+    OrganizationForm: "AS",
+    Status: "Active",
+    ParentOrganizationNumber: "",
+  },
+  {
+    Name: "Krever innlogging",
+    Type: "Business",
+    OrganizationNumber: "999999998",
+    OrganizationForm: "BEDR",
+    Status: "Active",
+    ParentOrganizationNumber: "311111111",
+  },
+  {
+    Name: "Generell feil",
+    Type: "Business",
+    OrganizationNumber: "999999997",
+    OrganizationForm: "BEDR",
+    Status: "Active",
+    ParentOrganizationNumber: "311111111",
+  },
+];
